requirements.txt
lxml==4.9.3
python-docx==1.0.1
typing-extensions==4.8.0
